chore: reintroducing statsd
